Move to package "updatecheck"
